Share tests between ApplyBQSRIntegrationTest and Spark equivalent.
